bug fix -"//TODO მთლიანი ციკლის დამთავრებისას ბოლოში აგდებს ექსეპშენს, ვერ გავარკვიე რისია" - გამოსწორდა.   ბოლო აპდეითის შემდეგ გმირის დაღუპვის შემთხვევაში თამაში ჩერდებოდა და არ გამოდიოდა ციკლიდან - გამოსწორდა.   გასწორდა რამდენიმე "TIPO" შეცდომა.   გალამაზდა კოდის დიზაინი

+++ src/MyGame/Random/Random.kt
@@ -1,16 +1,12 @@
 package MyGame.Random
 
-import MyGame.Character.MainChar
-import MyGame.Units.Enemy
-import MyGame.Units.Player
-
 /**
  * ღებულობს ორ ცვლადს და რანდომ აბრუნებს რომელიმეს.
  */
-fun WhoStart(player: String, enemy: String):String{
+fun WhoStart(player: String, enemy: String): String {
     var starter: String = "enemy"
-    val i  = (0..1).random()
-    if(i == 1){
+    val i = (0..1).random()
+    if (i == 1) {
         starter = "player"
     }
     return starter
@@ -28,10 +24,10 @@ fun WhoStart(player: String, enemy: String):String{
  * აბრუნებს თრუ ორ ფალსს, რის მიხედვითაც ვსაზღვრავთ, დაარტყა თუ ააცილა
  *  //TODO უნიტს უნდა დაემატოს acuracy, რომელიც განსაზღვრავს დარტყმის ძალის პროცენტს
  */
-fun Hit_Or_No():Boolean{
+fun Hit_Or_No(): Boolean {
     var hit = false
     val int = (0..10).random()
-    if(int > 3){
+    if (int > 3) {
         hit = true
     }
     return hit
@@ -41,8 +37,9 @@ fun Hit_Or_No():Boolean{
  *  პრიმიტიული ლოგიკით ვსაზღვრავთ, რა ძალითდ აარტყა
  *  //TODO მომავალში უნდა გადმოეცეს უნიტის დარტყმის ძალა(maxDemage-ში) და იქედან განისაზღვროს მნიშვნელობა
  */
-fun Hit_Point(maxDemage: Int): Int{
+fun Hit_Point(maxDemage: Double): Double {
     val minDemage = maxDemage / 10
-    val hitPoint = (minDemage..maxDemage).random()
-    return hitPoint
+    //TODO ინტეჯერები გადასაყვანია დაბლებში, რომ დააბრუნოს უფრო ზუსტი მნიშვნელობა ჰიტპოინტისთვის
+    val hitPoint = (minDemage.toInt()..maxDemage.toInt()).random()
+    return hitPoint.toDouble()
 }
